% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_landsat.R
\name{get_scene_table}
\alias{get_scene_table}
\title{Generate table of Landsat scenes}
\usage{
get_scene_table()
}
\value{
data.frame containing Landsat scene attributes from
"https://landsat-pds.s3.amazonaws.com/c1/L8/scene_list.gz"
}
\description{
Generate table of Landsat scenes
}
