% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_landsat.R
\name{merge_rasts}
\alias{merge_rasts}
\title{Build multiband raster from several .tifs}
\usage{
merge_rasts(out_file, in_files)
}
\arguments{
\item{out_file}{file path to .tif file to be created}

\item{in_files}{file paths of .tif files to be added to .vrt}
}
\value{
file path to output file
}
\description{
Build multiband raster from several .tifs
}
