% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_landsat.R
\name{resample_raster}
\alias{resample_raster}
\title{Build multiband raster from several .tifs}
\usage{
resample_raster(file, a_nodata, tr)
}
\arguments{
\item{file}{file path to .tif file to be resampled}

\item{a_nodata}{nodata value for output raster}

\item{tr}{target resolution}
}
\value{
file path to output .tif file
}
\description{
Build multiband raster from several .tifs
}
