% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_landsat.R
\name{view_landsat}
\alias{view_landsat}
\title{Preview Landsat 8 images with `mapview`}
\usage{
view_landsat(p, r, date)
}
\arguments{
\item{p}{landsat path (integer)}

\item{r}{landsat row (integer)}

\item{date}{image date formatted as YYYYMMDD}
}
\value{
`mapview` interactive map
}
\description{
For a specified scene create a `mapview` interactive map with a false color composite
image (bands 5, 4, 3), the cirrus band (band 9), and the QA band
}
